Refuse new connections during graceful shutdown
Unlike std::net::TCPListener, tokio::net::TCPListener will internally
accept TCP connections even when the application has stopped calling
accept on the TCPListener.
As such, connection attempts during a graceful shutdown will be
accepted by tokio, but will not be processed; once the shutdown
completes, the connections will be closed.

This commit drops the TCPListener before waiting for the shutdown
to complete, so that new connection attempts during a shutdown are
immediately refused.